codegen(llvm): rename `context.rs` into `ctx.rs`
